requirements.txt
# requirements.txt
Django>=5.0.0,<6.0.0
Pillow>=10.0.0,<11.0.0
requests>=2.30.0,<3.0.0
python-docx>=1.0.0,<2.0.0
python-dotenv>=1.0.0,<2.0.0
